requirements.txt
aiogram==3.10.0
Flask==3.0.3
python-dotenv==1.0.1
aiohttp~=3.9.5
aiosqlite==0.20.0
uvicorn==0.30.1
asgiref==3.8.1
requests==2.32.3
